Consistent sentinels for event-logging IDs, encounter types, symptoms
Event-logging sentinel/fallback choices had drifted: bare -1/255 literals
repeated with no named constant, and infector_symptom_id used 0 both as
the real "recovered" registry id and as Disease::getSymptomId's own
error fallback, so seed-infection rows (no real infector) were
indistinguishable from a recovered infector.

- Add kInvalidPersonId/kInvalidVenueId (-1), kUnknownEncounterTypeId (255),
  kNoSymptomId (255) in core/types.h; replace bare literals across
  PendingInfection, PersonLocation, config.h encounter-type fields.
- Narrow infector_symptom_id/old_symptom_id/new_symptom_id uint16_t -> uint8_t
  end to end (event structs, HDF5 CompType registrations, logger API),
  freeing symptom id 0 to mean only "recovered". Disease's internal
  uint16_t representation is untouched; narrowing happens only at the
  logInfection/logSymptomChange call boundary via explicit static_cast.
- Standardise "nothing to report" string fallbacks in lookup tables:
  "unknown" -> "could not resolve" (value existed but couldn't be
  resolved), "" -> "not applicable" (nothing to report).

Verified: full build passes, test_integration and test_simulator both
pass, and a standalone EventLogger spot-check confirms the HDF5 output
now has uint8 symptom-id columns with 255 (not 0) for seed infections.

+++ include/utils/event_logging/event_types.h
@@ -35,7 +35,7 @@ struct InfectionEvent {
   double time;
   uint8_t encounter_type_id;
   uint8_t transmission_mode_index;
-  uint16_t infector_symptom_id;
+  uint8_t infector_symptom_id;
   InfectionSource source;
   uint8_t reserved = 0;
 };
@@ -44,8 +44,8 @@ struct SymptomChangeEvent {
   PersonId person_id;
   VenueId venue_id;
   double time;
-  uint16_t old_symptom_id;
-  uint16_t new_symptom_id;
+  uint8_t old_symptom_id;
+  uint8_t new_symptom_id;
 };
 
 struct DeathEvent {
@@ -123,8 +123,8 @@ struct SymptomChangeRecord {
   int person_id;
   int venue_id;
   double time;
-  uint16_t old_symptom_id;
-  uint16_t new_symptom_id;
+  uint8_t old_symptom_id;
+  uint8_t new_symptom_id;
 };
 
 struct HospitalAdmissionRecord {
